fix(money): migrate AppointmentSlotDTO.price to integer minor-units
price field changed from double to int. fromJson uses .round() to handle
Postgres numeric type. All display sites divide by 100. service_form_modal
converts user input (major) to minor on save. parseMoneyMinor(service.price)
call sites replaced with service.price directly. DB rows pre-migrated via
UPDATE appointment_slots SET price = ROUND(price*100) WHERE price < 10000.

Hardcoded GHS currency label in manage_services_screen replaced with draft
currency from shopCreationProvider.

+++ lib/presentation/features/shops/dashboard/presentation/screens/pricing_overrides_list_screen.dart
@@ -266,7 +266,7 @@ class _ErrorState extends StatelessWidget {
 
 class _OverrideRow extends StatelessWidget {
   final PricingOverrideDTO rule;
-  final double basePrice;
+  final int basePrice; // minor units
   final VoidCallback onEdit;
   final VoidCallback onArchive;
